+++ app/backend/app/Models/RolePermissions.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use App\Models\Permissions;
+use App\Models\Masters\Permissions;
 use App\Models\Roles;
 
 class RolePermissions extends Model
