+++ app/backend/database/seeders/masters/AdminsRolesTableSeeder.php
@@ -1,19 +1,22 @@
 <?php
 
-namespace Database\Seeders;
+namespace Database\Seeders\Masters;
 
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Config;
-use App\Models\Permissions;
+use App\Models\Masters\AdminsRoles;
 use Database\Seeders\BaseSeeder;
 
-class PermissionsTableSeeder extends BaseSeeder
+class AdminsRolesTableSeeder extends BaseSeeder
 {
-    protected const SEEDER_DATA_LENGTH = 4;
-    protected const SEEDER_DATA_TESTING_LENGTH = 4;
-    protected const SEEDER_DATA_DEVELOP_LENGTH = 4;
+    protected const SEEDER_DATA_LENGTH = 5;
+    protected const SEEDER_DATA_TESTING_LENGTH = 5;
+    protected const SEEDER_DATA_DEVELOP_LENGTH = 50;
+
+    // 一番権限の小さいロールのID
+    private const MINIMUM_ROLE_ID = 5;
 
     /**
      * Run the database seeds.
@@ -22,17 +25,19 @@ class PermissionsTableSeeder extends BaseSeeder
      */
     public function run()
     {
-        $this->tableName = (new Permissions())->getTable();
+        $this->tableName = (new AdminsRoles())->getTable();
 
         $now = Carbon::now()->timezone(Config::get('app.timezone'));
 
         $template = [
-            'name'       => '',
+            'admin_id'   => 1,
+            'role_id'    => 1,
             'created_at' => $now,
             'updated_at' => $now
         ];
 
-        $dataList = Config::get('myappSeeder.seeder.authority.permissionsNameList');
+        // insert用データ
+        $data = [];
 
         // データ数
         $this->count = $this->getSeederDataLengthByEnv(
@@ -42,14 +47,12 @@ class PermissionsTableSeeder extends BaseSeeder
             self::SEEDER_DATA_DEVELOP_LENGTH
         );
 
-        // insert用データ
-        $data = [];
-
         // 1~$this->countの数字の配列でforを回す
         foreach (range(1, $this->count) as $i) {
             $row = $template;
 
-            $row['name'] = $dataList[$i - 1];
+            $row['admin_id'] = $i;
+            $row['role_id']  = $i > self::MINIMUM_ROLE_ID ? self::MINIMUM_ROLE_ID : $i;
 
             $data[] = $row;
         }
